requirements.txt
folium==0.20.0
haversine==2.9.0
numpy==2.3.1
pandas==2.3.1
Pillow==11.3.0
plotly==6.2.0
streamlit==1.47.0
